admin-end/requirements.txt
Flask
Flask-Bootstrap
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
pandas
matplotlib
gunicorn
python-dotenv
flask-wtf
PyPDF2
python-docx